refactor: 검증 및 예외처리 관련 메서드를 묶어서 별도의 파일로 분리
